Add ability to have cell array parameters that get recorded in NEV
It's a little clunky, but this version of sending cell info to the NEV
can only handle single depth cells (i.e. not cells-in-cells) and works
by writing out a line of code that can be read out from the NEV for each
entry to the cell. I feel like there should be a nicer form to do it all
in 'one line' that might be faster to transfer to the NEV, but for the
moment it only happens once per session (or block changee), I think, so
I'm not too worried about it being clunky.